+++ generators/typescript/express/cli/src/custom-config/ExpressCustomConfig.ts
@@ -15,5 +15,4 @@ export interface ExpressCustomConfig {
     outputSourceFiles: boolean;
     retainOriginalCasing: boolean;
     allowExtraFields: boolean;
-    useJsFileExtensionsOnImports: boolean;
 }