+++ generators/typescript/utils/commons/src/referencing/getRelativePathAsModuleSpecifierTo.ts
@@ -5,18 +5,12 @@ export function getRelativePathAsModuleSpecifierTo({
     from,
     to,
     isDesinationADirectory,
-    useJsExtension,
-    appendBarrelFileForDirectoryImport = useJsExtension
+    isOutputtingEsm
 }: {
     from: Directory | SourceFile | string;
     to: Directory | SourceFile | string;
     isDesinationADirectory: boolean;
-    useJsExtension: boolean;
-    /**
-     * whether to add `index.js` to the relative path.
-     * defaults to the value of `useJsExtension`
-     */
-    appendBarrelFileForDirectoryImport?: boolean;
+    isOutputtingEsm: boolean;
 }): string {
     const parsedToFilePath = path.parse(getPath(to));
     const toFilePathWithoutExtension = path.join(parsedToFilePath.dir, parsedToFilePath.name);
@@ -32,10 +26,10 @@ export function getRelativePathAsModuleSpecifierTo({
     if (moduleSpecifier.endsWith("/")) {
         moduleSpecifier = moduleSpecifier.slice(0, -1);
     }
-    if (isDesinationADirectory && appendBarrelFileForDirectoryImport) {
-        moduleSpecifier += "/index";
-    }
-    if (useJsExtension) {
+    if (isOutputtingEsm) {
+        if (isDesinationADirectory) {
+            moduleSpecifier += "/index";
+        }
         moduleSpecifier += ".js";
     }
     return moduleSpecifier;