+++ seed/ts-express/exhaustive/js-file-extensions/api/resources/endpoints/resources/index.ts
@@ -1,7 +1,0 @@
-export * as container from "./container/index.js";
-export * as enum_ from "./enum/index.js";
-export * as httpMethods from "./httpMethods/index.js";
-export * as object from "./object/index.js";
-export * as params from "./params/index.js";
-export * as primitive from "./primitive/index.js";
-export * as union from "./union/index.js";