+++ seed/ts-express/exhaustive/js-file-extensions/index.ts
@@ -1,3 +1,0 @@
-export * as SeedExhaustive from "./api/index.js";
-export { register } from "./register.js";
-export { SeedExhaustiveError } from "./errors/index.js";