+++ seed/ts-express/exhaustive/js-file-extensions/serialization/resources/endpoints/resources/httpMethods/service/index.ts
@@ -1,2 +1,0 @@
-export * as testGet from "./testGet.js";
-export * as testDelete from "./testDelete.js";