+++ seed/ts-express/exhaustive/js-file-extensions/serialization/resources/endpoints/resources/params/service/index.ts
@@ -1,2 +1,0 @@
-export * as getWithPath from "./getWithPath.js";
-export * as modifyWithPath from "./modifyWithPath.js";