+++ seed/ts-express/exhaustive/js-file-extensions/serialization/resources/types/resources/index.ts
@@ -1,6 +1,0 @@
-export * as enum_ from "./enum/index.js";
-export * from "./enum/types/index.js";
-export * as object from "./object/index.js";
-export * from "./object/types/index.js";
-export * as union from "./union/index.js";
-export * from "./union/types/index.js";